+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/platform.kt
@@ -9,3 +9,5 @@ internal expect inline fun AtomicBoolean.set(value: Boolean)
 internal expect inline fun AtomicBoolean.compareAndSet(expect: Boolean, update: Boolean): Boolean
 
 internal expect inline fun atomicBooleanOf(initialValue: Boolean): AtomicBoolean
+
+internal expect inline fun nanoTime(): Long